requirements.txt
streamlit>=1.28.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=1.0.0
langchain-huggingface>=0.0.1
langchain-chroma>=0.1.0
langgraph>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pymupdf>=1.23.0
python-docx>=1.1.0
tqdm>=4.65.0
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0
pypdf>=3.0.0
langchain-text-splitters>=0.2.0
langchain-core>=0.2.0

